tidy up: remove /images and move cash.js to /lib
