--- .coveralls.yml
@@ -1,1 +1,0 @@
-repo_token: L78Rmx7kRcKgWQcPAo82esSxcqBdMY21l
