+++ lib/change_calculator.js
