+++ test/change_calculator.test.js
@@ -1,5 +1,5 @@
 var test = require('tape'); // assing the tape library to the variable "test"
-var C = require('../cash.js');  // our module
+var C = require('../lib/change_calculator.js');  // our module
 
 test('should return -1 when the value is not present', function (t) {
   t.equal(-1, [1,2,3].indexOf(4)); // 4 is not present in this array so
